DAY 3/enterprise_leave_mgmt/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
pydantic[email]
pytest
pytest-asyncio
httpx
alembic